refactored to support new device model
